Add support for native Python deps.
All native deps are linked together into a statically-linked shared object; the
modules themselves are dynamically linked to it. The actual native modules are
compiled separately.

This is similar to the internal Google model, as far as I understand.

A detail is that we always alwayslink the native deps. Theoretically it
should've been enough to alwayslink only the direct dependencies of toplevel
modules: the linker can't know which of their symbols are used by the toplevel
modules when linking the nativedeps SO. In practice, I've seen cases where this
caused missed symbol errors. Haven't been able to reproduce this in small-scale
examples, and BUILD files did seem correct.

Tested on both Linux and macOS.