Add support for native Python deps.
All native deps are linked together into a statically-linked shared object; the
modules themselves are dynamically linked to it, which causes it to be loaded,
making available the PyInit_* functions.

This is similar to the internal Google model, except (if I understood correctly)
they compile the extension code itself into the module SO, and we keep it empty.

Tested on both Linux and macOS.